feat: split unit from integration tests in jest config
