+++ dev-client/__tests__/integration/CreateSiteScreen-test.tsx
@@ -15,7 +15,7 @@
  * along with this program. If not, see https://www.gnu.org/licenses/.
  */
 
-import {render} from '@testing/utils';
+import {render} from '@testing/snapshot/utils';
 
 import {CreateSiteScreen} from 'terraso-mobile-client/screens/CreateSiteScreen/CreateSiteScreen';
